test(loading): migrate trigger
